requirements.txt
accelerate>=1.13.0,<2.0.0
Flask>=3.1.0,<4.0.0
huggingface_hub>=1.8.0,<2.0.0
playwright>=1.52.0,<2.0.0
pypdf>=6.0.0,<7.0.0
python-dotenv>=1.0.1,<2.0.0
python-docx>=1.2.0,<2.0.0
requests>=2.32.0,<3.0.0
safetensors>=0.7.0,<1.0.0
sentencepiece>=0.2.1,<1.0.0
soundfile>=0.13.1,<1.0.0
torch>=2.11.0,<3.0.0
torchaudio>=2.11.0,<3.0.0
transformers>=5.4.0,<6.0.0
